validate and double check the following:
Default Plotted Variables to 2: In the "Weather & Marine Data Explorer" (/om-marine-explorer/page.tsx), I'll update the initialVisibility useMemo hook so that only the first two parameters in ALL_PARAMETERS are set to true by default. The plotVisibility state will then be initialized with this, ensuring only two plots are visible when the page first loads or data is fetched.
X-axis Time Format to "dd/MM/yy": In the ChartDisplay.tsx component, I'll simplify the memoizedXAxisTickFormatter and formatDateTickBrush functions to always format the time as "dd/MM/yy". The dynamic formatting based on time range duration will be removed.